Refactor app to solo user experience, remove multi-role support
