+++ src/components/Analytics/Readiness/TableView.tsx
@@ -42,7 +42,7 @@ export const TableView: React.FC<TableViewProps> = ({ data, period, isLoading })
 
       {/* ARIA Insights */}
       <section className="card">
-        <h2 className="text-lg font-semibold mb-2">Coach ARIA Readiness Insights</h2>
+        <h2 className="text-lg font-semibold mb-2">ARIA Readiness Insights</h2>
         <p className="text-sm text-gray-600 mb-4">
           AI-generated insights about your readiness trends
         </p>
