+++ src/components/ProtectedRoute.tsx
@@ -5,11 +5,10 @@ import { useAuth } from '@/contexts/AuthContext';
 
 interface ProtectedRouteProps {
   children: React.ReactNode;
-  roles?: ('coach' | 'athlete' | 'solo')[];
 }
 
-const ProtectedRoute: React.FC<ProtectedRouteProps> = ({ children, roles }) => {
-  const { session, profile, loading } = useAuth();
+const ProtectedRoute: React.FC<ProtectedRouteProps> = ({ children }) => {
+  const { session, loading } = useAuth();
 
   if (loading) {
     return (
@@ -24,11 +23,6 @@ const ProtectedRoute: React.FC<ProtectedRouteProps> = ({ children, roles }) => {
 
   if (!session) {
     return <Navigate to="/home" replace />;
-  }
-
-  // Check role-based access if roles are specified
-  if (roles && profile && !roles.includes(profile.role)) {
-    return <Navigate to="/dashboard" replace />;
   }
 
   return <>{children}</>;
