+++ src/pages/LoadDetailPage.tsx
@@ -6,7 +6,7 @@ import { InsightStrip } from '@/components/Analytics/InsightStrip';
 import { SegmentedControl } from '@/components/Analytics/SegmentedControl';
 import { AnalyticsHeroSection } from '@/components/Analytics/AnalyticsHeroSection';
 import { useMetricData } from '@/hooks/useMetricData';
-import { useEnhancedMetricsWithAthlete } from '@/hooks/useEnhancedMetricsWithAthlete';
+import { useEnhancedMetrics } from '@/hooks/useEnhancedMetrics';
 import { Dumbbell } from 'lucide-react';
 import { MuscleAnatomyView } from '@/components/Analytics/MuscleAnatomyView';
 
@@ -35,7 +35,7 @@ export default function LoadDetailPage() {
   const { data, isLoading, error } = useMetricData("load", period);
 
   // Get enhanced metrics for InsightStrip
-  const { readinessRolling, sleepDaily, loadACWR, latestStrain } = useEnhancedMetricsWithAthlete(profile?.id);
+  const { readinessRolling, sleepDaily, loadACWR, latestStrain } = useEnhancedMetrics();
 
   // Extract values for InsightStrip
   const latestReadiness = readinessRolling[readinessRolling.length - 1]?.readiness_score ?? null;
