+++ src/pages/Sleep.tsx
@@ -6,7 +6,7 @@ import { PeriodProvider } from '@/lib/hooks/usePeriod';
 import { InsightStrip } from '@/components/Analytics/InsightStrip';
 import { AnalyticsHeroSection } from '@/components/Analytics/AnalyticsHeroSection';
 import { SegmentedControl } from '@/components/Analytics/SegmentedControl';
-import { useEnhancedMetricsWithAthlete } from '@/hooks/useEnhancedMetricsWithAthlete';
+import { useEnhancedMetrics } from '@/hooks/useEnhancedMetrics';
 import { useAuth } from '@/contexts/AuthContext';
 import { ArrowLeft, Moon } from 'lucide-react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
@@ -34,7 +34,7 @@ const SleepContent: React.FC = () => {
   const avgSleepHours = getAverageSleepHours();
 
   // Get enhanced metrics for InsightStrip
-  const { readinessRolling, sleepDaily, loadACWR, latestStrain } = useEnhancedMetricsWithAthlete(profile?.id);
+  const { readinessRolling, sleepDaily, loadACWR, latestStrain } = useEnhancedMetrics();
 
   // Extract values for InsightStrip
   const latestReadiness = readinessRolling[readinessRolling.length - 1]?.readiness_score ?? null;
